Moved Base64OutputStreamFactory to http package
BugId: 48522
